# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Leet-Pi API
  description: |
    Backend escalable para MVP de evaluación de ejercicios de programación en C.

    ## Autenticación
    La mayoría de los endpoints requieren un JWT Bearer token.  
    Obtené el token haciendo `POST /api/auth/login` con un token Auth0 válido en el header `Authorization: Bearer <auth0_token>`.  
    El servidor retorna el JWT de la plataforma en el header de respuesta **`X-Auth-Token`**. Usá ese valor para llamadas autenticadas subsecuentes.

    ## Roles
    - **student** – puede enviar soluciones y ver sus propias submissions.
    - **admin** – acceso a operaciones administrativas de guías, ejercicios y usuarios normales.
    - **superadmin** – nivel más alto: todo lo que puede admin más gestión de admins y asignación de roles.
  version: 1.0.0
  contact:
    name: ITBA – leet-pi team
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por la plataforma (`X-Auth-Token`). Obtenido vía `POST /api/auth/login`.
    Auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT emitido directamente por Auth0. Requerido únicamente en `POST /api/auth/login`.

  # ─── Common response schemas ───────────────
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          description: Payload variado según el endpoint
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            statusCode:
              type: integer
          required:
            - message
            - statusCode
      required:
        - success
        - error

    # ─── Domain schemas ──────────────────────
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        sub:
          type: string
          description: Auth0 subject identifier
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        submissionCount:
          type: integer
...
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: User not authenticated
              statusCode: 401
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: Obtener perfil del usuario autenticado
      description: Sincroniza y retorna el perfil completo del usuario a partir de su token Auth0.
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil obtenido
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
              example:
                success: true
                message: User profile retrieved successfully
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  email: user@example.com
                  firstName: Juan
                  lastName: Pérez
                  sub: auth0|abc123
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me:
    patch:
      tags: [Users]
      summary: Actualizar nombre y apellido del usuario autenticado
      description: Permite al usuario autenticado actualizar su firstName y lastName.
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName]
              properties:
                firstName:
                  type: string
                  example: Juan
                lastName:
                  type: string
                  example: Pérez
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          email:
                            type: string
                          firstName:
                            type: string
                          lastName:
                            type: string
              example:
                success: true
                message: Profile updated successfully
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  email: user@example.com
                  firstName: Juan
                  lastName: Pérez
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Users]
      summary: Información básica del usuario actual
      description: Retorna id, email, firstName y lastName del usuario autenticado.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Usuario actual obtenido
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          email:
                            type: string
                          firstName:
                            type: string
                          lastName:
                            type: string
              example:
                success: true
                message: Current user retrieved successfully
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  email: user@example.com
                  firstName: Juan
                  lastName: Pérez
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /users:
    get:
      tags: [Users]
      summary: Listar todos los usuarios
      description: Retorna todos los usuarios registrados. Requiere autenticación (uso administrativo).
      operationId: getAllUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserSummary'
              example:
                success: true
                message: Users retrieved successfully
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    email: user@example.com
                    firstName: Juan
                    lastName: Pérez
                    submissionCount: 5
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}:
    get:
      tags: [Users]
      summary: Obtener usuario por ID
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Users]
      summary: Eliminar usuario por ID
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario a eliminar
      responses:
        '200':
          description: Usuario eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: User deleted successfully
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ═══════════════════════════════════════════
  # SUBMISSIONS
  # ═══════════════════════════════════════════
...